Specify the host so the API always finds the release.
There was an issue where since I'm authenticated to an enterprise
Github, this refused to find the release. Even running the switch with
`gh auth switch` and switching to github.com didn't resolve the issue.
However, since this command's documentation shows host is an expected,
optional component of the `--repo` argument, and we know this should be
looking there, this avoids any trouble. Now the build and install works
correctly.