fix(config.lua): Remove warning about using Copilot for auto_suggestions
For those who would like to or have to use GitHub Copilot a constant notification every time they open neovim might be too annoying.
I've added the same warning that's in the README.md to the config.lua. This should be enough of a warning to anyone that decides to use GitHub Copilot for auto_suggestions.