fix: don't rebalance windows after opening/closing the sidebar
The `wincmd =` call in `Sidebar:open()` overrides the widths/heights of
the sidebar windows that were previously set. This only happens in
certain situations though for some reason, e.g. when using the
horizontal layout.

Either way rebalancing seems unnecessary and is also intrusive since it
rearranges all windows, not just those the plugin uses.
The call in `Sidebar:close()` is removed for the same reason.