fix: update displayed response for the first chunk (#1594)
If one is using avante.nvim with a non streaming LLM API the API
response will only have a single chunk. So that if we don't update the
displayed_response in the sidebar component, the output of the LLM will
be removed from the sidebar when we complete the interaction with the
LLM.

This commit makes sure that the displayed_response is updated for the
first chunk as well as subsequent ones.