improve: resolve AWS credentials using default credentials provider chain for bedrock
This switches to resolving AWS credentials using the default AWS
credentials provider chain.

This has the advantage that AWS credentials are resolved in the same way
as other AWS tools do. It also supports more mechanisms now like AWS
profiles, single-sign on and more.

It is a breaking change for bedrock users, as the old `BEDROCK_KEYS` env
var will no longer work, however it gives much more flexibility in terms
of authentication.