feat: add config for ollama keep alive
the default timeout (5m) isn't large enough for my use case. i would
rather prefer to keep the model in memory for longer to avoid the
process of reloading it.

this commit introduces the ollama `keep_alive` configuration, the
default is set to 5 minutes (that is the default for ollama).