fix: apply re-uses the correct buffer when using non-cwd project root
The buffer name is the relative path to vim.fn.getcwd() which is not
necessarily the same as the project root. However, we were using this to
prepend the filepath to the buffer name and match up the filepath being
applied to the relevant buffer. This resulted in a bug where apply
opened an unnamed buffer in the background instead of the relevant
buffer. Thus we use vim.fn.getcwd() to prepend the buffer name which
makes the buffer for the filepath properly being re-used.