fix: properly highlight multi filetype code blocks
Code blocks for files with multiple filetypes are tagged with the entire
multiple filetype combination. For example, yaml.github has a codeblock
with yaml.github tagged in the codeblock, when yaml should be used. This
updates the codeblock to only use the first filetype (and possibly fixes
other usages of `get_filetype` as well).