optimize(selection): Improve visual mode hint autocommand management
Triggering autocommands on CursorMove event is pretty expensive,
especially if they are not buffer-local, as they are triggered very
often. Additionally the command itself is more expensive as it has to
test whether it was triggered for the right buffer and right mode.

Refactor the visual mode autocommands to be created for particular
buffer when entering visual mode, and cleaned up upon exiting visual
mode:

- there is only global "ModeChanged" autocommand that triggers when
  entering visual mode
- a new autocommand group is created when entering visual mode and
  "CursorMoved", second "ModeChanged", and "BufLeave" autocommands are
  made buffer-local and moved there
- the new group is deleted when leaving visual mode or leaving buffer

Making most of the autocommands local to the buffer reduces unnecessary
checks and improves the overall efficiency of hint display.