optimize(sidebar): Optimize token counting for displaying in input hint
As the size of context and prompt grows so grows the cost to tokenize
the prompt and calculate the number of tokens. Given that we try to
update "input hint" containing the estimated number of tokens on each
text change, and we do it synchronously, this causes major delays when
typing in Avante's input buffer.

Fix the problem by calculating and caching the number of tokens in the
prompt on submission (and also when restoring an old chat history) and
then only calculating the number of tokens in the input buffer which is
much smaller and therefore much faster.

This also simplifies Sidebar:show_input_hint() as it no longer users
multiple callbacks but instead executes sequentially.

Fixes #711