optimize(sidebar): use debounce when updating hint due to text change
Updating input hint (and therefore removing the window, recalculating
tokens for the input portion of the prompt, and rendering window again)
is wasteful if done on each "TextChanged" event as it happens pretty
often.

Use debouncer so that the input hint is updated only when typing is
paused for 200 msec or more. This reduces wasteful work while still
presenting reasonably up-to-date token counts.