optimize(sidebar): consolidate ModeChange handling for input hint
Contrary to the code comment we do not hide input hint when leaving
insert mode, we simply update it because submit key sequence is likely
to be different (<CR> vs <C-s> by default).

Consolidate the 2 autocommand into one and tie it to the input buffer
now that we do not need to match on the mode pattern (which does not
work together with buffer number option).