fix(llm_tools): Improve the permissions logic
I've had an issue where the plugin thinks it doesn't have permissions
within a folder.

This is partly due to the fact that my home directory is under version
control, and some of the directories I work within are in .gitignore.

This fix seems to resolve the issue, which I found in one of the issues,
and decided to add the fix to the repo.

This seems to be an edge case, but equally ... I'm sure I'm not the only
one faced with this issue

This seems to be an edge case, but equally ... I'm sure I'm not the only
one faced with this issue

Signed-off-by: Yomi Colledge <baphled@boodah.net>