Add config to limit max height for selected files
This commit adds a configuration option to set a maximum height for the
selected files window. This is because at the moment, if you attached a
lot of selected files, it would end up taking the whole chat buffer
which made the user experience really rough.

After adding this configuration option, I just modified the sidebar
logic so that instead of deciding between the minimum of the maximum
number of height lines in the vim window, or the number of selected
files, we are now picking between the configured max lines height, and
the number of selected files, which should make the UX much better.