fix(sidebar): fix closing sidebar when closing its components
We want to close entire sidebar when user closes its input or results
window. This functionality was broken when we stopped tracking window
IDs of windows belonging to the sidebar separately from containers
(NuiSplit) that represent them. This happens because Both sidebar and
NuiSplit register WinClosed handlers, and NuiSplit ones fire first,
resetting window data in the container so that
Sidebar:get_sidebar_window() no longer works.

Fix this by making sure that Sidebar's WinClosed is registered first so
that it is executed before NuiSplit resets its window ID to nil.