fix(sidebar): avoid crashing in WinNew handler
If "WinNew" autocommand set up in Sidebar:setup_colors() fires at an
inopportune time it will reference self.code.winid and error out with
"Invalid window ID".

Fix this by adding check to test if code window is valid before
attempting to use it. This should fix with the stack trace in #2770.